+++ src/App.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Provider } from 'jotai';
 import Game from './components/Game';
 import ThemeToggle from './components/ThemeToggle';
@@ -7,21 +7,17 @@ import InstallPrompt from './components/InstallPrompt';
 import UpdateNotification from './components/UpdateNotification';
 import AnalyticsProvider from './components/AnalyticsProvider';
 import LanguageSelector from './components/LanguageSelector';
-import SettingsDialog from './components/SettingsDialog';
+import SettingsButton from './components/settings/SettingsButton';
 import { LanguageProvider } from './hooks/useLanguage';
 import { useLanguage } from './hooks/useLanguage';
 import './App.scss';
 import './styles/DarkTheme.scss';
-import './styles/SettingsDialog.scss';
+import './components/settings/styles/SettingsDialog.scss';
 import '@fortawesome/fontawesome-free/css/all.css';
 
 // Wrap the app content in the LanguageContext
 const AppContent: React.FC = () => {
   const { t } = useLanguage();
-  const [isSettingsOpen, setIsSettingsOpen] = useState(false);
-
-  const openSettings = () => setIsSettingsOpen(true);
-  const closeSettings = () => setIsSettingsOpen(false);
 
   return (
     <div className="App">
@@ -30,13 +26,7 @@ const AppContent: React.FC = () => {
         <div className="header-right">
           <LanguageSelector />
           <ThemeToggle />
-          <button 
-            className="settings-button" 
-            onClick={openSettings}
-            aria-label="Open settings"
-          >
-            <i className="fas fa-cog settings-icon"></i>
-          </button>
+          <SettingsButton />
         </div>
       </header>
       <main>
@@ -54,7 +44,6 @@ const AppContent: React.FC = () => {
       <OfflineNotice />
       <InstallPrompt />
       <UpdateNotification />
-      <SettingsDialog isOpen={isSettingsOpen} onClose={closeSettings} />
     </div>
   );
 };
