+++ src/components/settings/styles/SettingsDialog.scss
@@ -1,4 +1,4 @@
-@use 'index' as *;
+@use '../../../styles/index' as *;
 
 .overlay {
   @include overlay-dark;
